requirements.txt
Flask==2.2.5
flask-cors==3.0.10
gunicorn==20.1.0
transformers==4.36.2
tokenizers==0.15.0
torch==2.1.0
openai==1.13.3
python-dotenv==1.0.1

